Autocancel futures
future variable which cancels all obeserves when last reference to it is destroyed - normal future remains alive to notifiy its observers, however, observes may be invalid because they refer to already destroyed object. So FutureAutoCancel can cancel all observers when it is destroyed, because it is probably destroyed with the object where the observers refers to.